standalone cigar & cigar_test
