+++ src/leviosam_utils.cpp
@@ -185,65 +185,6 @@ bool WriteDeferred::commit_aln_dest(const bam1_t* const aln) {
     return true;
 }
 
-void update_cigar(bam1_t* aln, std::vector<uint32_t>& new_cigar) {
-    uint32_t* cigar = bam_get_cigar(aln);
-    // Adapted from samtools/sam.c
-    // https://github.com/samtools/htslib/blob/2264113e5df1946210828e45d29c605915bd3733/sam.c#L515
-    if (aln->core.n_cigar != new_cigar.size()) {
-        auto cigar_st = (uint8_t*)bam_get_cigar(aln) - aln->data;
-        auto fake_bytes = aln->core.n_cigar * 4;
-        aln->core.n_cigar = (uint32_t)new_cigar.size();
-        auto n_cigar4 = aln->core.n_cigar * 4;
-        auto orig_len = aln->l_data;
-        if (n_cigar4 > fake_bytes) {
-            // Check if we need to update `aln->m_data`.
-            //
-            // Adapted from htslib/sam_internal.h
-            // https://github.com/samtools/htslib/blob/31f0a76d338c9bf3a6893b71dd437ef5fcaaea0e/sam_internal.h#L48
-            auto new_m_data = (size_t)aln->l_data + n_cigar4 - fake_bytes;
-            kroundup32(new_m_data);
-            if (new_m_data > aln->m_data) {
-                auto new_data =
-                    static_cast<uint8_t*>(realloc(aln->data, new_m_data));
-                if (!new_data) {
-                    std::cerr << "[Error] Failed to expand a bam1_t struct for "
-                              << bam_get_qname(aln) << "\n";
-                    std::cerr << "This is likely due to out of memory\n";
-                    exit(1);
-                }
-                aln->data = new_data;
-                aln->m_data = new_m_data;
-                cigar = bam_get_cigar(aln);
-            }
-        }
-        // Update memory usage of data.
-        aln->l_data = aln->l_data - fake_bytes + n_cigar4;
-        // Move data to the correct place.
-        memmove(aln->data + cigar_st + n_cigar4,
-                aln->data + cigar_st + fake_bytes,
-                orig_len - (cigar_st + fake_bytes));
-        // If new n_cigar is greater, copy the real CIGAR to the right place.
-        // Skipped this if new n_cigar is smaller than the original value.
-        if (n_cigar4 > fake_bytes) {
-            memcpy(aln->data + cigar_st,
-                   aln->data + (n_cigar4 - fake_bytes) + 8, n_cigar4);
-        }
-        aln->core.n_cigar = new_cigar.size();
-    }
-    for (int i = 0; i < aln->core.n_cigar; i++) {
-        *(cigar + i) = new_cigar[i];
-    }
-}
-
-void debug_print_cigar(uint32_t* cigar, size_t n_cigar) {
-    for (int i = 0; i < n_cigar; i++) {
-        auto cigar_op_len = bam_cigar_oplen(cigar[i]);
-        auto cigar_op = bam_cigar_op(cigar[i]);
-        std::cerr << cigar_op_len << bam_cigar_opchr(cigar_op);
-    }
-    std::cerr << "\n";
-}
-
 /** Removes the MN:i and MD:z tags from an alignment object.
  *
  * @param aln Alignment object.
@@ -318,12 +259,17 @@ std::string get_read(const bam1_t* rec) {
     return read;
 }
 
-/* Update SAM flag to set a record as an unmapped alignment
+/**
+ * Sets an alignment to unmapped.
  *   - Clear forward/reverse status.
  *   - If paired, changed to improper paired.
  *
- * We currenly set an unliftable read as unampped, and thus the
- * BAM_FUNMAP or BAM_FMUNMAP flags will be raised.
+ * An unliftable read is considered as unmapped. The BAM_FUNMAP or BAM_FMUNMAP
+ * flags are updated accordingly.
+ *
+ * @param aln
+ * @param first_set
+ * @param keep_mapq
  */
 void update_flag_unmap(bam1_t* aln, const bool first_seg,
                        const bool keep_mapq = false) {
